Made some comments
I now see the problem of putting 
Let's say, you analyse data and you send the output to shortstack/
Then, if my script takes as input the shortstack/ folder and list all folders in there (to retrieve sample names for instance), then it will also take I have a script that list the directory.
To make our life easier when we share the same input data, we should keep the input folder and files as "read only". 
Let's work on that.